Fix possible rounding errors of music/sfx volume at startup

+++ src/c_cmds.c
@@ -4206,9 +4206,8 @@ static void s_volume_cvars_func2(char *cmd, char *parm1, char *parm2, char *parm
         if (M_StringCompare(cmd, stringize(s_musicvolume)) && s_musicvolume != value)
         {
             s_musicvolume = value;
-            musicVolume = (BETWEEN(s_musicvolume_min, s_musicvolume,
-                s_musicvolume_max) * 15 + 50) / 100;
-            S_SetMusicVolume((int)(musicVolume * 127 / 15.0));
+            musicVolume = (BETWEEN(s_musicvolume_min, s_musicvolume, s_musicvolume_max) * 15 + 50) / 100;
+            S_SetMusicVolume(musicVolume * 127 / 15);
 
             M_SaveCVARs();
         }
@@ -4216,7 +4215,7 @@ static void s_volume_cvars_func2(char *cmd, char *parm1, char *parm2, char *parm
         {
             s_sfxvolume = value;
             sfxVolume = (BETWEEN(s_sfxvolume_min, s_sfxvolume, s_sfxvolume_max) * 15 + 50) / 100;
-            S_SetSfxVolume((int)(sfxVolume * 127 / 15.0));
+            S_SetSfxVolume(sfxVolume * 127 / 15);
 
             M_SaveCVARs();
         }
